requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# API & Validation
pydantic==2.4.2
python-multipart==0.0.6
bcrypt==4.1.1
pyjwt==2.8.1

# Database
sqlalchemy==2.0.23

# Utilities
python-dateutil==2.8.2
pytz==2023.3
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# API & Validation
pydantic==2.4.2
python-multipart==0.0.6
bcrypt==4.1.1
pyjwt==2.8.1

# Database
sqlalchemy==2.0.23

# Utilities
python-dateutil==2.8.2
pytz==2023.3
